(multiplayer-basic) Update clientId to a 64 bit uint to allow unique identification of every IP address and port.  Also cleaned up a few other related item.

+++ C++/MultiPlayer/Step 1 - Basic/server/GameModel.hpp
@@ -18,18 +18,18 @@ class GameModel
     void shutdown();
 
   private:
-    std::unordered_set<std::uint32_t> m_players;
-    std::unordered_map<std::uint32_t, entities::Entity::IdType> m_clientIdToEntityId;
+    std::unordered_set<std::uint64_t> m_players;
+    std::unordered_map<std::uint64_t, entities::Entity::IdType> m_clientIdToEntityId;
     entities::EntityMap m_entities;
     std::unordered_set<entities::Entity::IdType> m_reportThese;
 
     std::unique_ptr<systems::Network> m_systemNetwork;
 
-    void clientConnected(std::uint32_t clientId);
+    void clientConnected(std::uint64_t clientId);
     void addEntity(std::shared_ptr<entities::Entity> entity);
     void removeEntity(entities::Entity::IdType entityId);
 
     void updateClients();
-    void handleJoin(std::uint32_t clientId, std::shared_ptr<messages::Join> message);
-    void handleInput(std::uint32_t clientId, std::shared_ptr<messages::Input> message);
+    void handleJoin(std::uint64_t clientId, std::shared_ptr<messages::Join> message);
+    void handleInput(std::uint64_t clientId, std::shared_ptr<messages::Input> message);
 };
